+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -36,6 +36,7 @@
 * Addition of `evalTxExUnits` and `evalTxExUnitsWithLogs`. Deprecated
   `evaluateTransactionExecutionUnits` and `evaluateTransactionExecutionUnitsWithLogs` if
   favor of the newly added ones.
+* Export `hashDataTxWitsL`
 
 ## 1.0.0.0